Fix input glitches
