+++ src/components/S/DefaultArray.vue
