+++ src/components/S/DefaultString.vue
